+++ config/detekt/detekt.yml
@@ -128,7 +128,7 @@ complexity:
     threshold: 600
   LongMethod:
     active: true
-    threshold: 60
+    threshold: 75
   LongParameterList:
     active: true
     functionThreshold: 6
